// client/src/redux/general/general.types.ts
export enum GeneralTypes {
  SHOW_ALERT = "SHOW_ALERT",
  HIDE_ALERT = "HIDE_ALERT",
  SHOW_FULL_SIZE_SEARCHBOX = "SHOW_FULL_SIZE_SEARCHBOX",
  HIDE_FULL_SIZE_SEARCHBOX = "HIDE_FULL_SIZE_SEARCHBOX",
  SHOW_SEARCH_PANEL = "SHOW_SEARCH_PANEL",
  HIDE_SEARCH_PANEL = "HIDE_SEARCH_PANEL",
  OPEN_DIALOG = "OPEN_DIALOG",
  CLOSE_DIALOG = "CLOSE_DIALOG",
  OPEN_FULL_SIZE_PHOTO_VIEWER = "OPEN_FULL_SIZE_PHOTO_VIEWER",
  CLOSE_FULL_SIZE_PHOTO_VIEWER = "CLOSE_FULL_SIZE_PHOTO_VIEWER",
  TOGGLE_DARK_THEME = "TOGGLE_DARK_THEME",
  SET_DARK_THEME_ENABLED = "SET_DARK_THEME_ENABLED",
}